fix: use the new `configuredFor` method in all places
A know bug is that when you have another ALTERNATIVE configured
execution, the email OTP will always show in the list of "other ways" to login.
It will not execute or be required, but it will be shown.

This is because internal keycloak classes use the "normal" `configuredFor` method
which does not take into consideration the flow configuration.

These uses were spotted in (✅ means fixed, either in this pr or in main & ❌ means not yet fixed):
- ✅ https://github.com/keycloak/keycloak/blob/579b185e7af60d741c97a1c73e975506b0c1e56c/services/src/main/java/org/keycloak/authentication/authenticators/conditional/ConditionalUserConfiguredAuthenticator.java#L61
- ✅ https://github.com/keycloak/keycloak/blob/579b185e7af60d741c97a1c73e975506b0c1e56c/services/src/main/java/org/keycloak/forms/login/freemarker/AuthenticatorConfiguredMethod.java#L46
- ❌ https://github.com/keycloak/keycloak/blob/579b185e7af60d741c97a1c73e975506b0c1e56c/services/src/main/java/org/keycloak/authentication/DefaultAuthenticationFlow.java#L458
- ❌ https://github.com/keycloak/keycloak/blob/579b185e7af60d741c97a1c73e975506b0c1e56c/services/src/main/java/org/keycloak/authentication/FormAuthenticationFlow.java#L197